# src/main/java/com/example/bicicletario/bicicletario/domain/models/Bicicleta.java
package com.example.bicicletario.bicicletario.domain.models;

import com.example.bicicletario.bicicletario.domain.enums.StatusBicicleta;

public class Bicicleta {

    private Integer id;

    private String marca;

    private String modelo;

    private String ano;

    private Integer numero;

    private StatusBicicleta statusBicicleta;

    private String dataInsercaoTranca;

    private String dataRemocaoTranca;

    private Integer idFuncionarioUltimaOperacao;

    public Bicicleta() {
    }

    public Bicicleta(String marca, String modelo, String ano, Integer numero, StatusBicicleta statusBicicleta) {
        this.marca = marca;
        this.modelo = modelo;
        this.ano = ano;
        this.numero = numero;
        this.statusBicicleta = statusBicicleta;
    }

    // Getters e Setters

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getMarca() {
        return marca;
    }

    public void setMarca(String marca) {
        this.marca = marca;
    }

    public String getModelo() {
        return modelo;
    }

    public void setModelo(String modelo) {
        this.modelo = modelo;
    }

    public String getAno() {
        return ano;
    }

    public void setAno(String ano) {
        this.ano = ano;
    }

    public Integer getNumero() {
        return numero;
    }

    public void setNumero(Integer numero) {
        this.numero = numero;
    }

    public StatusBicicleta getStatusBicicleta() {
        return statusBicicleta;
    }

    public void setStatusBicicleta(StatusBicicleta statusBicicleta) {
        this.statusBicicleta = statusBicicleta;
    }

    public String getDataInsercaoTranca() {
        return dataInsercaoTranca;
    }

    public void setDataInsercaoTranca(String dataInsercaoTranca) {
        this.dataInsercaoTranca = dataInsercaoTranca;
    }

    public String getDataRemocaoTranca() {
        return dataRemocaoTranca;
    }

    public void setDataRemocaoTranca(String dataRemocaoTranca) {
        this.dataRemocaoTranca = dataRemocaoTranca;
    }

    public Integer getIdFuncionarioUltimaOperacao() {
        return idFuncionarioUltimaOperacao;
    }

    public void setIdFuncionarioUltimaOperacao(Integer idFuncionarioUltimaOperacao) {
        this.idFuncionarioUltimaOperacao = idFuncionarioUltimaOperacao;
    }
}