# src/main/java/com/example/bicicletario/bicicletario/domain/models/Totem.java
package com.example.bicicletario.bicicletario.domain.models;

public class Totem {

    private Integer id;

    private String localizacao;

    private String descricao;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getLocalizacao() {
        return localizacao;
    }

    public void setLocalizacao(String localizacao) {
        this.localizacao = localizacao;
    }

    public String getDescricao() {
        return descricao;
    }

    public void setDescricao(String descricao) {
        this.descricao = descricao;
    }
}