# src/main/resources/static/swagger.yaml
openapi: 3.0.0
info:
  title: Sistema de controle de bicicletário (29/04/2023)
  description: Modelo da disciplina. Documentação relacionada ao sistema de controle de bicicletário. Toda a API está documentada neste arquivo de definição.
  contact:
    email: pasemes@uniriotec.br
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1"
servers:
- url: http://localhost:8020/api
  description: SwaggerHub API Auto Mocking
...
paths:
...
  /bicicleta:
    get:
      tags:
      - Equipamento
      summary: recupera bicicletas cadastradas
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bicicleta'
    post:
      tags:
      - Equipamento
      summary: Cadastrar bicicleta
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovaBicicleta'
        required: true
      responses:
        "200":
          description: Dados cadastrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bicicleta'
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
  /bicicleta/integrarNaRede:
    post:
      tags:
      - Equipamento
      summary: colocar uma bicicleta nova ou retornando de reparo de volta na rede de totens
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bicicleta_integrarNaRede_body'
        required: true
      responses:
        "200":
          description: Dados cadastrados
        "422":
          description: Dados Inválidos (ex status inválido da bicicleta ou tranca)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
  /bicicleta/retirarDaRede:
    post:
      tags:
      - Equipamento
      summary: retirar bicicleta para reparo ou aposentadoria
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bicicleta_retirarDaRede_body'
        required: true
      responses:
        "200":
          description: Dados cadastrados
        "422":
          description: Dados Inválidos (ex status inválido da bicicleta ou tranca)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
  /bicicleta/{idBicicleta}:
    get:
      tags:
      - Equipamento
      summary: Obter bicicleta
      parameters:
      - name: idBicicleta
        in: path
        description: UUID da bicicleta
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Bicicleta encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bicicleta'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
    put:
      tags:
      - Equipamento
      summary: Editar bicicleta
      parameters:
      - name: idBicicleta
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovaBicicleta'
        required: true
      responses:
        "200":
          description: Dados atualizados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bicicleta'
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
    delete:
      tags:
      - Equipamento
      summary: Remover bicicleta
      parameters:
      - name: idBicicleta
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Dados removidos
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /bicicleta/{idBicicleta}/status/{acao}:
    post:
      tags:
      - Equipamento
      summary: Alterar status da bicicleta
      parameters:
      - name: idBicicleta
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: acao
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: "enum['DISPONIVEL','EM_USO', 'NOVA', 'APOSENTADA', 'REPARO_SOLICITADO', 'EM_REPARO' ]"
      responses:
        "200":
          description: Ação bem sucedida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bicicleta'
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /totem:
    get:
      tags:
      - Equipamento
      summary: recupera totens cadastrados
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Totem'
    post:
      tags:
      - Equipamento
      summary: Incluir totem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/novoTotem'
        required: true
      responses:
        "200":
          description: Dados cadastrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Totem'
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
...
  /totem/{idTotem}:
    put:
      tags:
      - Equipamento
      summary: Editar totem
      parameters:
      - name: idTotem
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/novoTotem'
        required: true
      responses:
        "200":
          description: Dados cadastrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Totem'
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
    delete:
      tags:
      - Equipamento
      summary: Remover totem
      parameters:
      - name: idTotem
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Totem removido
        "404":
          description: Totem não encontrado
  /totem/{idTotem}/trancas:
    get:
      tags:
      - Equipamento
      summary: Listar trancas de um totem
      parameters:
      - name: idTotem
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tranca'
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /totem/{idTotem}/bicicletas:
    get:
      tags:
      - Equipamento
      summary: Listar bicicletas de um totem
      parameters:
      - name: idTotem
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bicicleta'
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...
components:
  schemas:
...
    NovaBicicleta:
      type: object
      properties:
        marca:
          type: string
        modelo:
          type: string
        ano:
          type: string
          format: year
        numero:
          type: integer
        status:
          type: string
          format: "enum['DISPONÍVEL','EM_USO', 'NOVA', 'APOSENTADA', 'REPARO_SOLICITADO', 'EM_REPARO' ]"
    Bicicleta:
      allOf:
      - type: object
        properties:
          id:
            type: integer
      - $ref: '#/components/schemas/NovaBicicleta'
    novoTotem:
      required:
      - localizacao
      type: object
      properties:
        localizacao:
          type: string
          format: coordinates
        descricao:
          type: string
          format: string
    Totem:
      allOf:
      - type: object
        properties:
          id:
            type: integer
      - $ref: '#/components/schemas/novoTotem'
    NovaTranca:
      type: object
      properties:
        numero:
          minimum: 0
          exclusiveMinimum: true
          type: integer
        localizacao:
          type: string
          format: coordinates
        anoDeFabricacao:
          type: string
          format: year
        modelo:
          type: string
        status:
          type: string
          format: "enum['LIVRE', 'OCUPADA', 'NOVA', 'APOSENTADA', 'EM_REPARO']"
    Tranca:
      allOf:
      - type: object
        properties:
          id:
            type: integer
          bicicleta:
            type: integer
      - $ref: '#/components/schemas/NovaTranca'
...
    Erro:
      required:
      - codigo
      - mensagem
      type: object
      properties:
        codigo:
          type: string
        mensagem:
          type: string
...
    bicicleta_integrarNaRede_body:
      type: object
      properties:
        idTranca:
          type: integer
        idBicicleta:
          type: integer
        idFuncionario:
          type: integer
    bicicleta_retirarDaRede_body:
      type: object
      properties:
        idTranca:
          type: integer
        idBicicleta:
          type: integer
        idFuncionario:
          type: integer
        statusAcaoReparador:
          type: string
          description: "Status escolhido como destino do equipamento, uma vez que o reparo tenha sido solicitado anteriormente."
          format: "enum['APOSENTADA', 'EM_REPARO' ]"
...